+++ apps/api/src/auth/jwt/decodeJWT.ts
@@ -1,6 +1,0 @@
-import { JWT } from "@phylopic/source-models"
-import { decode, JwtPayload } from "jsonwebtoken"
-const decodeJWT = (token: JWT) => {
-    return decode(token, { json: true }) as JwtPayload | null
-}
-export default decodeJWT